change API from events to exposing isDayHighlighted
