+++ packages/components/src/date-time/date.js
@@ -24,7 +24,6 @@ class DatePicker extends Component {
 		this.onChangeMoment = this.onChangeMoment.bind( this );
 		this.nodeRef = createRef();
 		this.keepFocusInside = this.keepFocusInside.bind( this );
-		this.isDayHighlighted = this.isDayHighlighted.bind( this );
 	}
 
 	/*
@@ -82,23 +81,8 @@ class DatePicker extends Component {
 		return currentDate ? moment( currentDate ) : moment();
 	}
 
-	isDayHighlighted( date ) {
-		const { events } = this.props;
-		if ( ! events?.length ) {
-			return;
-		}
-		return events.some( ( event ) => date.isSame( event.date, 'day' ) );
-	}
-
-	componentWillReceiveProps( nextProps ) {
-		// Changing the reference to isDayHighlighted makes the calendar recompute highlighted days.
-		if ( nextProps.events !== this.props.events ) {
-			this.isDayHighlighted = this.isDayHighlighted.bind( this );
-		}
-	}
-
 	render() {
-		const { currentDate, isInvalidDate } = this.props;
+		const { currentDate, isInvalidDate, isDayHighlighted } = this.props;
 
 		const momentDate = this.getMomentDate( currentDate );
 
@@ -123,7 +107,7 @@ class DatePicker extends Component {
 					isOutsideRange={ ( date ) => {
 						return isInvalidDate && isInvalidDate( date.toDate() );
 					} }
-					isDayHighlighted={ this.isDayHighlighted }
+					isDayHighlighted={ isDayHighlighted }
 					onPrevMonthClick={ this.keepFocusInside }
 					onNextMonthClick={ this.keepFocusInside }
 				/>
