+++ packages/components/src/date-time/index.js
@@ -42,8 +42,8 @@ export class DateTimePicker extends Component {
 			currentDate,
 			is12Hour,
 			isInvalidDate,
+			isDayHighlighted,
 			onChange,
-			events,
 		} = this.props;
 
 		return (
@@ -59,7 +59,7 @@ export class DateTimePicker extends Component {
 							currentDate={ currentDate }
 							onChange={ onChange }
 							isInvalidDate={ isInvalidDate }
-							events={ events }
+							isDayHighlighted={ isDayHighlighted }
 						/>
 					</>
 				) }
